test: split report OOM tests
On some machines the report OOM tests can take too long to complete,
resulting in a timeout. This splits the test into several different
smaller tests to reduce the flakiness of it.
